+++ rana_qgis_plugin/widgets/rana_browser.py
@@ -390,9 +390,11 @@ class FilesBrowser(QWidget):
         self.files_tv.header().setSortIndicatorShown(True)
         self.files_tv.doubleClicked.connect(self.select_file_or_directory)
         self.btn_upload = QPushButton("Upload Files to Rana")
+        self.btn_new_schematisation = QPushButton("New schematisation")
         layout = QVBoxLayout(self)
         layout.addWidget(self.files_tv)
         layout.addWidget(self.btn_upload)
+        layout.addWidget(self.btn_new_schematisation)
         self.setLayout(layout)
 
     def refresh(self):
@@ -820,6 +822,7 @@ class RanaBrowser(QWidget):
     create_model_selected_with_revision = pyqtSignal(dict, int)
     open_schematisation_selected_with_revision = pyqtSignal(dict, dict)
     delete_file_selected = pyqtSignal(dict, dict)
+    upload_new_schematisation_selected = pyqtSignal(dict)
 
     def __init__(self, communication: UICommunication):
         super().__init__()
@@ -949,6 +952,10 @@ class RanaBrowser(QWidget):
             file_browser_signal.connect(
                 lambda file, signal=rana_signal: signal.emit(self.project, file)
             )
+        # Connect new schematisation button
+        self.files_browser.btn_new_schematisation.clicked.connect(
+            lambda _,: self.upload_new_schematisation_selected.emit(self.project)
+        )
         # Connect updating folder from breadcrumb
         self.breadcrumbs.folder_selected.connect(
             lambda path: self.files_browser.select_path(path)